refactor(view): extract InputView from Application
